Move part of `redemption.ts` to the `./services/redemptions` feature module
Move the user-facing redemption code to the new `./services/redemptions` module.
Some refactoring meant to facilitate usage was done as well.
