+++ typescript/src/services/redemptions/index.ts
@@ -1,0 +1,1 @@
+export * from "./redemptions-service"
